+++ core/src/server/helpers.rs
@@ -24,67 +24,13 @@
 // IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use std::borrow::Cow as StdCow;
 use std::io;
 use std::time::Duration;
 
-use jsonrpsee_types::error::{
-	reject_too_big_batch_response, ErrorCode, ErrorObject, OVERSIZED_RESPONSE_CODE, OVERSIZED_RESPONSE_MSG,
-};
-use jsonrpsee_types::{ErrorObjectOwned, Id, InvalidRequest, Response, ResponsePayload};
-use serde::Serialize;
-use serde_json::value::to_raw_value;
+use jsonrpsee_types::{ErrorCode, ErrorObject, Id, InvalidRequest, Response, ResponsePayload};
 use tokio::sync::mpsc;
 
 use super::{DisconnectError, SendTimeoutError, SubscriptionMessage, TrySendError};
-use crate::server::LOG_TARGET;
-
-/// Wrapper over [`tokio::mpsc::oneshot::channel`]..
-pub fn response_channel() -> (MethodResponseNotifyTx, MethodResponseNotifyRx) {
-	let (tx, rx) = tokio::sync::oneshot::channel();
-	(tx, MethodResponseNotifyRx(rx))
-}
-
-/// ..
-pub type MethodResponseNotifyTx = tokio::sync::oneshot::Sender<NotifyMsg>;
-
-/// ..
-#[derive(Debug)]
-pub struct MethodResponseNotifyRx(tokio::sync::oneshot::Receiver<NotifyMsg>);
-
-/// A message that that tells whether notification
-/// was succesful or not.
-#[derive(Debug, Copy, Clone)]
-pub enum NotifyMsg {
-	/// The response was succesfully processed.
-	Ok,
-	/// The response was the wrong kind
-	/// such an error response when
-	/// one expected a succesful response.
-	WrongKind,
-}
-
-/// ...
-#[derive(Debug, Copy, Clone)]
-pub enum MethodResponseError {
-	/// The connection was closed.
-	Closed,
-	/// The response was the wrong kind
-	/// such an error response when
-	/// one expected a succesful response.
-	WrongKind,
-}
-
-impl MethodResponseNotifyRx {
-	/// Whether the message was sent or not.
-	pub async fn is_sent(self) -> Result<(), MethodResponseError> {
-		match self.0.await {
-			Ok(NotifyMsg::Ok) => Ok(()),
-			Ok(NotifyMsg::WrongKind) => Err(MethodResponseError::WrongKind),
-			Err(_) => Err(MethodResponseError::Closed),
-		}
-	}
-}
 
 /// Bounded writer that allows writing at most `max_len` bytes.
 ///
@@ -216,382 +162,10 @@ pub fn prepare_error(data: &[u8]) -> (Id<'_>, ErrorCode) {
 	}
 }
 
-#[derive(Debug, Clone)]
-pub(crate) enum ResponseKind {
-	MethodCall,
-	Subscription,
-	Batch,
-}
-
-/// Notification kind.
-#[derive(Debug)]
-pub enum NotifyKind {
-	/// Just notify on success.
-	Success(MethodResponseNotifyTx),
-	/// Just notify on error.
-	Error(MethodResponseNotifyTx),
-	/// Notify on both success or error.
-	All(MethodResponseNotifyTx),
-}
-
-/// Represents a response to a method call.
-///
-/// NOTE: A subscription is also a method call but it's
-/// possible determine whether a method response
-/// is "subscription" or "ordinary method call"
-/// by calling [`MethodResponse::is_subscription`]
-#[derive(Debug)]
-pub struct MethodResponse {
-	/// Serialized JSON-RPC response,
-	pub(crate) result: String,
-	/// Indicates whether the call was successful or not.
-	pub(crate) success_or_error: MethodResponseResult,
-	/// Indicates whether the call was a subscription response.
-	pub(crate) kind: ResponseKind,
-	/// Optional callback that may be utilized to notif
-	/// that the method response has been processed
-	pub(crate) on_close: Option<NotifyKind>,
-}
-
-impl MethodResponse {
-	/// Returns whether the call was successful.
-	pub fn is_success(&self) -> bool {
-		self.success_or_error.is_success()
-	}
-
-	/// Returns whether the call failed.
-	pub fn is_error(&self) -> bool {
-		self.success_or_error.is_success()
-	}
-
-	/// Returns whether the response is a subscription response.
-	pub fn is_subscription(&self) -> bool {
-		matches!(self.kind, ResponseKind::Subscription)
-	}
-
-	/// Returns whether the response is a method response.
-	pub fn is_method_call(&self) -> bool {
-		matches!(self.kind, ResponseKind::MethodCall)
-	}
-
-	/// Returns whether the response is a batch response.
-	pub fn is_batch(&self) -> bool {
-		matches!(self.kind, ResponseKind::Batch)
-	}
-}
-
-/// Represent the outcome of a method call success or failed.
-#[derive(Debug, Copy, Clone)]
-pub enum MethodResponseResult {
-	/// The method call was successful.
-	Success,
-	/// The method call failed with error code.
-	Failed(i32),
-}
-
-impl MethodResponseResult {
-	/// Returns whether the call was successful.
-	pub fn is_success(&self) -> bool {
-		matches!(self, MethodResponseResult::Success)
-	}
-
-	/// Returns whether the call failed.
-	pub fn is_error(&self) -> bool {
-		matches!(self, MethodResponseResult::Failed(_))
-	}
-
-	/// Get the error code
-	///
-	/// Returns `Some(error code)` if the call failed.
-	pub fn as_error_code(&self) -> Option<i32> {
-		match self {
-			Self::Failed(e) => Some(*e),
-			_ => None,
-		}
-	}
-}
-
-impl MethodResponse {
-	/// Consume the method response and extract the serialized response.
-	pub fn into_result(self) -> String {
-		self.result
-	}
-
-	/// Consume the method response and extract the parts.
-	pub fn into_parts(self) -> (String, Option<NotifyKind>) {
-		(self.result, self.on_close)
-	}
-
-	/// Get a reference to the serialized response.
-	pub fn as_result(&self) -> &str {
-		&self.result
-	}
-
-	/// Create a method response from [`BatchResponse`].
-	pub fn from_batch(batch: BatchResponse) -> Self {
-		Self {
-			result: batch.0,
-			success_or_error: MethodResponseResult::Success,
-			kind: ResponseKind::Batch,
-			on_close: None,
-		}
-	}
-
-	/// This is similar to [`MethodResponse::response`] but sets a flag to indicate
-	/// that response is a subscription.
-	pub fn subscription_response<T>(id: Id, result: ResponsePayloadV2<T>, max_response_size: usize) -> Self
-	where
-		T: Serialize + Clone,
-	{
-		let mut rp = Self::response(id, result, max_response_size);
-		rp.kind = ResponseKind::Subscription;
-		rp
-	}
-
-	/// Optional callback that may be utilized to notify that the method response was
-	/// a succesfully JSON-RPC object and processed.
-	pub fn notify_on_success(mut self, tx: MethodResponseNotifyTx) -> Self {
-		self.on_close = Some(NotifyKind::Success(tx));
-		self
-	}
-
-	/// Optional callback that may be utilized to notify that the method response was a
-	/// an JSON-RPC error object and processed.
-	pub fn notify_on_error(mut self, tx: MethodResponseNotifyTx) -> Self {
-		self.on_close = Some(NotifyKind::Error(tx));
-		self
-	}
-
-	/// Optional callback that may be utilized to notify that the method response was processed.
-	pub fn notify_on_response(mut self, tx: MethodResponseNotifyTx) -> Self {
-		self.on_close = Some(NotifyKind::All(tx));
-		self
-	}
-
-	/// Create a new method response.
-	///
-	/// If the serialization of `result` exceeds `max_response_size` then
-	/// the response is changed to an JSON-RPC error object.
-	pub fn response<T>(id: Id, rp: ResponsePayloadV2<T>, max_response_size: usize) -> Self
-	where
-		T: Serialize + Clone,
-	{
-		let mut writer = BoundedWriter::new(max_response_size);
-
-		let success_or_error = if let ResponsePayload::Error(ref e) = rp.inner {
-			MethodResponseResult::Failed(e.code())
-		} else {
-			MethodResponseResult::Success
-		};
-
-		let kind = ResponseKind::MethodCall;
-
-		match serde_json::to_writer(&mut writer, &Response::new(rp.inner, id.clone())) {
-			Ok(_) => {
-				// Safety - serde_json does not emit invalid UTF-8.
-				let result = unsafe { String::from_utf8_unchecked(writer.into_bytes()) };
-
-				Self { result, success_or_error, kind, on_close: rp.on_exit }
-			}
-			Err(err) => {
-				tracing::error!(target: LOG_TARGET, "Error serializing response: {:?}", err);
-
-				if err.is_io() {
-					let data = to_raw_value(&format!("Exceeded max limit of {max_response_size}")).ok();
-					let err_code = OVERSIZED_RESPONSE_CODE;
-
-					let err = ResponsePayload::error_borrowed(ErrorObject::borrowed(
-						err_code,
-						OVERSIZED_RESPONSE_MSG,
-						data.as_deref(),
-					));
-					let result =
-						serde_json::to_string(&Response::new(err, id)).expect("JSON serialization infallible; qed");
-
-					Self {
-						result,
-						success_or_error: MethodResponseResult::Failed(err_code),
-						kind,
-						on_close: rp.on_exit,
-					}
-				} else {
-					let err_code = ErrorCode::InternalError;
-					let result = serde_json::to_string(&Response::new(err_code.into(), id))
-						.expect("JSON serialization infallible; qed");
-					Self {
-						result,
-						success_or_error: MethodResponseResult::Failed(err_code.code()),
-						kind,
-						on_close: rp.on_exit,
-					}
-				}
-			}
-		}
-	}
-
-	/// This is similar to [`MethodResponse::error`] but sets a flag to indicate
-	/// that error is a subscription.
-	pub fn subscription_error<'a>(id: Id, err: impl Into<ErrorObject<'a>>) -> Self {
-		let mut rp = Self::error(id, err);
-		rp.kind = ResponseKind::Subscription;
-		rp
-	}
-
-	/// Create a [`MethodResponse`] from a JSON-RPC error.
-	pub fn error<'a>(id: Id, err: impl Into<ErrorObject<'a>>) -> Self {
-		let err: ErrorObject = err.into();
-		let err_code = err.code();
-		let err = ResponsePayload::error_borrowed(err);
-		let result = serde_json::to_string(&Response::new(err, id)).expect("JSON serialization infallible; qed");
-		Self {
-			result,
-			success_or_error: MethodResponseResult::Failed(err_code),
-			kind: ResponseKind::MethodCall,
-			on_close: None,
-		}
-	}
-}
-
-/// Builder to build a `BatchResponse`.
-#[derive(Debug, Clone, Default)]
-pub struct BatchResponseBuilder {
-	/// Serialized JSON-RPC response,
-	result: String,
-	/// Max limit for the batch
-	max_response_size: usize,
-}
-
-impl BatchResponseBuilder {
-	/// Create a new batch response builder with limit.
-	pub fn new_with_limit(limit: usize) -> Self {
-		let mut initial = String::with_capacity(2048);
-		initial.push('[');
-
-		Self { result: initial, max_response_size: limit }
-	}
-
-	/// Append a result from an individual method to the batch response.
-	///
-	/// Fails if the max limit is exceeded and returns to error response to
-	/// return early in order to not process method call responses which are thrown away anyway.
-	pub fn append(&mut self, response: &MethodResponse) -> Result<(), MethodResponse> {
-		// `,` will occupy one extra byte for each entry
-		// on the last item the `,` is replaced by `]`.
-		let len = response.result.len() + self.result.len() + 1;
-
-		if len > self.max_response_size {
-			Err(MethodResponse::error(Id::Null, reject_too_big_batch_response(self.max_response_size)))
-		} else {
-			self.result.push_str(&response.result);
-			self.result.push(',');
-			Ok(())
-		}
-	}
-
-	/// Check if the batch is empty.
-	pub fn is_empty(&self) -> bool {
-		self.result.len() <= 1
-	}
-
-	/// Finish the batch response
-	pub fn finish(mut self) -> BatchResponse {
-		if self.result.len() == 1 {
-			BatchResponse(batch_response_error(Id::Null, ErrorObject::from(ErrorCode::InvalidRequest)))
-		} else {
-			self.result.pop();
-			self.result.push(']');
-			BatchResponse(self.result)
-		}
-	}
-}
-
-/// Serialized batch response.
-#[derive(Debug, Clone)]
-pub struct BatchResponse(String);
-
-/// Create a JSON-RPC error response.
-pub fn batch_response_error(id: Id, err: impl Into<ErrorObject<'static>>) -> String {
-	let err = ResponsePayload::error_borrowed(err);
-	serde_json::to_string(&Response::new(err, id)).expect("JSON serialization infallible; qed")
-}
-
-/// Similar to [`ResponsePayload`] but possible to with an async-like
-/// API to detect when a method response has been sent.
-#[derive(Debug)]
-pub struct ResponsePayloadV2<'a, T>
-where
-	T: Clone,
-{
-	inner: ResponsePayload<'a, T>,
-	on_exit: Option<NotifyKind>,
-}
-
-impl<'a, T> ResponsePayloadV2<'a, T>
-where
-	T: Clone,
-{
-	/// Create a new [`ResponsePayloadNotify`].
-	pub fn new(inner: ResponsePayload<'a, T>) -> ResponsePayloadV2<'a, T> {
-		ResponsePayloadV2 { inner, on_exit: None }
-	}
-
-	/// Create successful an owned response payload.
-	pub fn result(t: T) -> Self {
-		ResponsePayloadV2::new(ResponsePayload::Result(StdCow::Owned(t)))
-	}
-
-	/// Create successful borrowed response payload.
-	pub fn result_borrowed(t: &'a T) -> Self {
-		ResponsePayloadV2::new(ResponsePayload::Result(StdCow::Borrowed(t)))
-	}
-
-	/// ..
-	pub fn notify_on_success(mut self, tx: MethodResponseNotifyTx) -> Self {
-		self.on_exit = Some(NotifyKind::Success(tx));
-		self
-	}
-
-	/// ..
-	pub fn notify_on_response(mut self, tx: MethodResponseNotifyTx) -> Self {
-		self.on_exit = Some(NotifyKind::All(tx));
-		self
-	}
-
-	/// Create successful partial response i.e, the `result field`
-	pub fn error(e: impl Into<ErrorObjectOwned>) -> Self {
-		let inner = ResponsePayload::Error(e.into());
-		Self::new(inner)
-	}
-
-	/// Create successful partial response i.e, the `result field`
-	pub fn error_borrowed(e: impl Into<ErrorObject<'a>>) -> Self {
-		let inner = ResponsePayload::Error(e.into());
-		Self::new(inner)
-	}
-
-	/// ..
-	pub fn notify_on_error(mut self, tx: MethodResponseNotifyTx) -> Self {
-		self.on_exit = Some(NotifyKind::Error(tx));
-		self
-	}
-}
-
-impl<'a, T> From<ErrorCode> for ResponsePayloadV2<'a, T>
-where
-    T: Clone
-{
-	fn from(code: ErrorCode) -> Self {
-		let inner = ResponsePayload::Error(code.into());
-		Self::new(inner)
-	}
-}
-
 #[cfg(test)]
 mod tests {
-	use crate::server::ResponsePayloadV2;
-
-	use super::{BatchResponseBuilder, BoundedWriter, Id, MethodResponse, Response};
-	use jsonrpsee_types::ResponsePayload;
+	use crate::server::BoundedWriter;
+	use jsonrpsee_types::{Id, Response, ResponsePayload};
 
 	#[test]
 	fn bounded_serializer_work() {
@@ -608,53 +182,5 @@ mod tests {
 		let mut writer = BoundedWriter::new(100);
 		// NOTE: `"` is part of the serialization so 101 characters.
 		assert!(serde_json::to_writer(&mut writer, &"x".repeat(99)).is_err());
-	}
-
-	#[test]
-	fn batch_with_single_works() {
-		let method = MethodResponse::response(Id::Number(1), ResponsePayloadV2::result_borrowed(&"a"), usize::MAX);
-		assert_eq!(method.result.len(), 37);
-
-		// Recall a batch appends two bytes for the `[]`.
-		let mut builder = BatchResponseBuilder::new_with_limit(39);
-		builder.append(&method).unwrap();
-		let batch = builder.finish();
-
-		assert_eq!(batch.0, r#"[{"jsonrpc":"2.0","result":"a","id":1}]"#)
-	}
-
-	#[test]
-	fn batch_with_multiple_works() {
-		let m1 = MethodResponse::response(Id::Number(1), ResponsePayloadV2::result_borrowed(&"a"), usize::MAX);
-		assert_eq!(m1.result.len(), 37);
-
-		// Recall a batch appends two bytes for the `[]` and one byte for `,` to append a method call.
-		// so it should be 2 + (37 * n) + (n-1)
-		let limit = 2 + (37 * 2) + 1;
-		let mut builder = BatchResponseBuilder::new_with_limit(limit);
-		builder.append(&m1).unwrap();
-		builder.append(&m1).unwrap();
-		let batch = builder.finish();
-
-		assert_eq!(batch.0, r#"[{"jsonrpc":"2.0","result":"a","id":1},{"jsonrpc":"2.0","result":"a","id":1}]"#)
-	}
-
-	#[test]
-	fn batch_empty_err() {
-		let batch = BatchResponseBuilder::new_with_limit(1024).finish();
-
-		let exp_err = r#"{"jsonrpc":"2.0","error":{"code":-32600,"message":"Invalid request"},"id":null}"#;
-		assert_eq!(batch.0, exp_err);
-	}
-
-	#[test]
-	fn batch_too_big() {
-		let method = MethodResponse::response(Id::Number(1), ResponsePayloadV2::result_borrowed(&"a".repeat(28)), 128);
-		assert_eq!(method.result.len(), 64);
-
-		let batch = BatchResponseBuilder::new_with_limit(63).append(&method).unwrap_err();
-
-		let exp_err = r#"{"jsonrpc":"2.0","error":{"code":-32011,"message":"The batch response was too large","data":"Exceeded max limit of 63"},"id":null}"#;
-		assert_eq!(batch.result, exp_err);
 	}
 }
