+++ server/src/transport/ws.rs
@@ -10,8 +10,7 @@ use futures_util::future::{self, Either};
 use futures_util::io::{BufReader, BufWriter};
 use futures_util::{Future, StreamExt, TryStreamExt};
 use hyper::upgrade::Upgraded;
-use jsonrpsee_core::server::helpers::{MethodSink, NotifyKind, NotifyMsg};
-use jsonrpsee_core::server::{BoundedSubscriptions, Methods};
+use jsonrpsee_core::server::{BoundedSubscriptions, MethodSink, Methods, NotifyKind, NotifyMsg};
 use jsonrpsee_types::error::{reject_too_big_request, ErrorCode};
 use jsonrpsee_types::Id;
 use soketto::connection::Error as SokettoError;
